Ignore experimental potions
